refactor: parse entity file path before displaying in charts and lists
