+++ apps/desktop/src/extensions/projects-view/ProjectDetails.tsx
@@ -14,7 +14,7 @@ import {
   DateRangeType,
   getRangeDates,
   mapRangeToPreset,
-} from "../../utils/dateRanges";
+} from "../../utils/time";
 import { addDays, startOfDay } from "date-fns";
 import { formatDuration } from "../../utils/time";
 import RangeSelectionDialog from "../../components/RangeSelectionDialog";
