+++ apps/desktop/src/utils/data.ts
@@ -1,0 +1,9 @@
+export const getEntityName = (
+  fullPath: string,
+  metadata: string | null,
+): string => {
+  if (metadata !== null && metadata === "File") {
+    return fullPath.split(/[\\/]/).pop() || fullPath;
+  }
+  return fullPath;
+};
